Migrate one GitLab structure to another instance:
Provided you edit the Global CONST variables at the top of the script,
you can query the structure of one GitLab instance and re-create it
on another. It creates all groups/projects the user has access to,
creating bare-bones repositories for each.

If there are projects you do not have privileges for, the script
will return an error, so it is advisable to log output e.g.

    yarn test >>test.log 2>&1